curium_surgai_backend/requirements.txt
paho-mqtt==2.1.0
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
numpy==1.26.4
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
requests==2.27.1
boto3==1.26.74
botocore==1.29.74
httpx==0.25.1
psycopg2-binary==2.9.5
psutil==5.9.1
drf-yasg==1.21.7